Commited changes

+++ PDF/pdf.php
@@ -20,16 +20,19 @@ class PDF extends FPDF
 
     public function __construct($formato, $mesini, $mesfin)
     {
+        include '../src/dbcred.php';
         $orientation='P'; $unit='mm'; $format='A4';
         parent::__construct($orientation,$unit,$format);
         $this->title = '';
-        $this->db = new PDO('pgsql:host=localhost;dbname=prometeo;', 'hans', 'prometeodb');
+        $this->db = $db;
         $this->nit = $_SESSION['nit'];
-        $this->formato = $formato;
-        $this->mesini = $mesini;
-        $this->mesfin = $mesfin;
+        $this->email = $_SESSION['email'];
+        $this->formato = (isset($formato))?$formato:null;
+        $this->mesini = (isset($mesini))?$mesini:null;
+        $this->mesfin = (isset($mesfin))?$mesfin:null;
         $this->docClass = new Doc($this->db);
-        $this->data_certificate = $this->data_certificate();
+        $this->data_certificate = ($this->data_certificate() != false)?$this->data_certificate():array();
+        $this->notaY = 165;
     }
     private function generar_id($type)
     {
@@ -42,7 +45,7 @@ class PDF extends FPDF
         $y = date('Y');
         $m = date('m');
         $d = date('d');
-        return $d.'/'.$m.'/'.$y;
+        return $m.'/'.$d.'/'.$y;
     }
     function doc_out($type='pdf')
     {
@@ -61,7 +64,9 @@ class PDF extends FPDF
                 $this->tabla1();
                 $this->tabla2();
                 $this->nota();
-                $this->Output();
+                $raw = $this->Output('S');
+                
+                return $raw;
             }else if(count($this->data_certificate) == 0){
                echo "Perdona la molestia, pero se a producido un error generando su certificado. El administrador ya ah sido notificado. Porfavor trate de nuevo mas tarder. Gracias por su paciencia.";
             }
@@ -76,19 +81,29 @@ class PDF extends FPDF
     }
     function data_certificate()
     {
-        $fields = $this->docClass->getFields($this->nit);
-        return $this->docClass->generar_certificado($this->formato, $this->mesini, $this->mesfin);
+        if($this->formato == null && $this->mesini == null && $this->mesfin == null)
+        {
+            return false;
+        }else{
+            $fields = $this->docClass->getFields($this->nit);
+            return $this->docClass->generar_certificado($this->formato, $this->mesini, $this->mesfin);
+        }
+        
     }
     function logo()
     {
-        $d = $_POST['img'];
-        $w = 20; $h = 25;
-        $raw = base64_decode($d);
-        $file = $_SERVER['DOCUMENT_ROOT'] . '/promTest/PDF/temp/'.rand() . '.jpeg';
-        $success = file_put_contents($file, $raw);
-        $this->Image($file,$w,$h);
-        unlink($file);
-    
+        if(isset($_POST['img']))
+        {
+            $d = $_POST['img'];
+            $w = 20; $h = 13;
+            $raw = base64_decode($d);
+            $file = $_SERVER['DOCUMENT_ROOT'] . '/promTest/PDF/temp/'.rand() . '.jpeg';
+            $success = file_put_contents($file, $raw);
+            $this->Image($file,$w,$h,50,40);
+            unlink($file);
+        }
+        
+        
     }
     function vars(){
         echo $this->title . '<br/>';
@@ -112,7 +127,6 @@ class PDF extends FPDF
     function guardar_certificado($folder_name)
     {
         $dir = $_SERVER['DOCUMENT_ROOT'] . '/promTest/PDF/documents/';
-        
         $path = (isset($folder_name) && $folder_name != null && $folder_name != 'null')?$dir.$this->nit.'/'.$folder_name:$dir.$this->nit;
         if(is_dir($path)) {
             $file = $this->data_certificate[0]['nmbre_frmto'].'_'.$this->generar_id('D');
@@ -130,7 +144,6 @@ class PDF extends FPDF
             $raw = $this->Output('S');
             $save = file_put_contents($path.'/'.$file.'.pdf', $raw);
             return $file.'.pdf';
-            
         }else{
             mkdir($path);
             $file = $this->data_certificate[0]['nmbre_frmto'].'_'.$this->generar_id('D');
@@ -148,7 +161,6 @@ class PDF extends FPDF
             $raw = $this->Output('S');
             $save = file_put_contents($path.'/'.$file.'.pdf', $raw);
             return $file.'.pdf';
-            
         }
     }
     
@@ -269,6 +281,7 @@ class PDF extends FPDF
            
         // Data
         $i = 0;
+        $n  = 0;
         $val1 = array();
         $val2 = array();
         foreach($data as $key => $row)
@@ -295,6 +308,12 @@ class PDF extends FPDF
                $i = 0;
                $this->Cell($headerCellWidth[$i],6,$row,0);
                $i++;
+               $n++;
+               if($n > 3)
+               {
+                 $this->notaY += 2;
+               }
+               
            }
         }
         if(count($data) > 4){
@@ -314,6 +333,7 @@ class PDF extends FPDF
         $this->SetFont('Arial','',10);
         $this->Cell(28,6,number_format($totalBase),'T');
         $this->Cell(28,6,number_format($totalRete),'T');
+        $this->Ln();
     }
     function nota()
     {
@@ -321,7 +341,7 @@ class PDF extends FPDF
         $pattern = '/(\d{2})\/(\d{2})\/(\d{4})/';
         preg_match($pattern, $subject, $matches, PREG_OFFSET_CAPTURE, 3);
         $txt = str_replace($matches[0][0], $this->fecha_de_hoy(), $subject);;
-        $this->SetY(145);
+        $this->SetY($this->notaY);
         $this->SetFont('Arial','',10);
         $this->SetTextColor(3,3,3);
         $this->MultiCell(0,5,$txt,0,'J',false);
@@ -367,6 +387,289 @@ class PDF extends FPDF
             return true;
         }
     }
+
+    function guardar_en_db($fname)
+    {
+        try{
+
+
+            $file = $this->data_certificate[0]['nmbre_rzon_scial'].'_'.$this->data_certificate[0]['nmbre_frmto'].'.pdf';
+            if($this->check_file($file, $fname) === false)
+            {
+                $this->SetMargins(20, 10, 20);
+                $this->AddPage();
+                $this->title = 'CERTIFICADO DE '.$this->data_certificate[0]['nmbre_frmto'].' ANO GRAVABLE '.substr($this->data_certificate[0]['ano_mes_fnal'],0,4);
+                $this->SetTitle($this->title);
+                $this->logo();
+                $this->Header();
+                $this->emDate();
+                $this->companyDetails();
+                $this->tabla1();
+                $this->tabla2();
+                $this->nota();
+    
+                $raw = pg_escape_bytea($this->Output('S'));
+    
+                $sql = "INSERT INTO usrios_web_mnjo_flders_det
+                (email_usrio,
+                nombre_folder,
+                asnto_dcmnto,
+                fcha_dcmnto,
+                nmbre_pdf )
+                VALUES ('".$this->email."',
+                '".$fname."',
+                '".$file."',
+                '".$this->fecha_de_hoy()."',
+                '".$raw."')";
+                $res = $this->db->query($sql);
+    
+                if($res)
+                {
+                    return $file;
+                }else{
+                    return false;
+                }
+
+            }else{
+                $pieces = explode('.', $file);
+                $file = $pieces[0].'_'.rand().'.'.$pieces[1];
+                $this->SetMargins(20, 10, 20);
+                $this->AddPage();
+                $this->title = 'CERTIFICADO DE '.$this->data_certificate[0]['nmbre_frmto'].' ANO GRAVABLE '.substr($this->data_certificate[0]['ano_mes_fnal'],0,4);
+                $this->SetTitle($this->title);
+                $this->logo();
+                $this->Header();
+                $this->emDate();
+                $this->companyDetails();
+                $this->tabla1();
+                $this->tabla2();
+                $this->nota();
+    
+                $raw = pg_escape_bytea($this->Output('S'));
+    
+                $sql = "INSERT INTO usrios_web_mnjo_flders_det
+                (email_usrio,
+                nombre_folder,
+                asnto_dcmnto,
+                fcha_dcmnto,
+                nmbre_pdf )
+                VALUES ('".$this->email."',
+                '".$fname."',
+                '".$file."',
+                '".$this->fecha_de_hoy()."',
+                '".$raw."')";
+                $res = $this->db->query($sql);
+    
+                if($res)
+                {
+                    return $file;
+                }else{
+                    return false;
+                }
+            }
+            
+        }catch(Exception $e){
+            return 'Err(guardar_en_db): '.  $e->getMessage(). "\n";
+        }
+        
+    }
+    function return_pdf($doc_id)
+    {
+        $sql = "SELECT nmbre_pdf FROM usrios_web_mnjo_flders_det WHERE email_usrio = '".$this->email."' AND id_dcmnto = ".$doc_id;
+        $res = $this->db->query($sql);
+        $row = $res->fetch(PDO::FETCH_ASSOC);
+        $result = pg_unescape_bytea($row['nmbre_pdf']);
+        return $result;
+    }
+    function all_pdf(){
+        $sql = "SELECT asnto_dcmnto FROM usrios_web_mnjo_flders_det WHERE email_usrio = '".$this->email."'";
+        $res = $this->db->query($sql);
+        $arr = array();
+        $row = $res->fetch(PDO::FETCH_ASSOC);
+        if(count($row) > 0)
+        {
+            while($row){
+                $arr['nombre'] = $row['asnto_dcmnto'].'.pdf';
+            }
+            return $arr;
+        }else{
+            $arr['nofiles'] = 'No hay documentos.';
+            return $arr;
+        }
+        
+    }
+    function check_user()
+    {
+        $sql = "SELECT COUNT(*) FROM usrios_web_mnjo_flders_enc WHERE email_usrio = '".$this->email."'";
+        $res = $this->db->query($sql);
+        $row = $res->fetch(PDO::FETCH_ASSOC);
+        if($row['count'] != 0)
+        {
+            return true;
+        }else{
+            return false;
+        }
+    }
+    function check_file($name, $fname)
+    {
+        $sql = "SELECT COUNT(*) FROM usrios_web_mnjo_flders_det WHERE asnto_dcmnto = '".$name."' AND email_usrio = '".$this->email."' AND nombre_folder = '".$fname."'";
+        $res = $this->db->query($sql);
+        $row = $res->fetch(PDO::FETCH_ASSOC);
+        if($row['count'] != 0)
+        {
+            return true;
+        }else{
+            return false;
+        }
+    }
+    function check_folder($name)
+    {
+        $sql = "SELECT COUNT(*) FROM usrios_web_mnjo_flders_enc WHERE nmbre_flder = '".$name."' AND email_usrio = '".$this->email."'";
+        $res = $this->db->query($sql);
+        $row = $res->fetch(PDO::FETCH_ASSOC);
+        if($row['count'] != 0)
+        {
+            return true;
+        }else{
+            return false;
+        }
+    }
+    function create_folder($name)
+    {
+        if($this->check_user() === false && isset($name))
+        {
+            $bnum = $this->generar_id('BE');
+
+            $sql = "INSERT INTO usrios_web_mnjo_flders_enc (email_usrio,
+            nmro_flder,
+            nmbre_flder,
+            nmro_orden)
+            VALUES('".$this->email."',
+            '".$bnum."',
+            'Bandeja de entrada',
+            1)";
+            $res = $this->db->query($sql);
+
+            if($res && $this->check_folder($name) !== true){
+                $fnum = $this->generar_id('F');
+                $sql = "INSERT INTO usrios_web_mnjo_flders_enc (email_usrio,
+                nmro_flder,
+                nmbre_flder)
+                VALUES('".$this->email."',
+                '".$fnum."',
+                '".$name."')";
+                $res = $this->db->query($sql);
+                if($res)
+                {
+                    return true;
+                }
+
+            }else if($this->check_folder($name) === true)
+            {
+                return 'Folder already exist';
+            }
+        }else if($this->check_user() === true && isset($name)){
+
+            if($this->check_folder($name) !== true){
+
+                $fnum = $this->generar_id('F');
+
+                $sql = "INSERT INTO usrios_web_mnjo_flders_enc (email_usrio,
+                nmro_flder,
+                nmbre_flder)
+                VALUES('".$this->email."',
+                '".$fnum."',
+                '".$name."')";
+                $res = $this->db->query($sql);
+
+                if($res)
+                {
+                    return true;
+                }
+
+            }else if($this->check_folder($name) === true)
+            {
+                return 'Folder already exist';
+            }
+           
+        }else{
+
+            return false;
+
+        }
+        
+       
+    }
+    function folder_list()
+    {
+        $sql = "SELECT * FROM usrios_web_mnjo_flders_enc WHERE email_usrio = '".$this->email."'";
+        $res = $this->db->query($sql);
+
+        $main = array();
+        $main['Bandeja de entrada'] = array();
+        $folder = array();
+        $i = 0;
+        while($row = $res->fetch(PDO::FETCH_ASSOC))
+        {
+            if($row['nmbre_flder'] == 'Bandeja de entrada')
+            {
+                $sqlfi = "SELECT * FROM usrios_web_mnjo_flders_det WHERE email_usrio = '".$this->email."' AND nombre_folder = 'Bandeja de entrada'";
+                $resfi = $this->db->query($sqlfi);
+                while($rowfi = $resfi->fetch(PDO::FETCH_ASSOC))
+                {
+                    
+                    $folder[$rowfi['asnto_dcmnto']] = $rowfi['id_dcmnto'];
+                }
+            }else{
+                $sqlf = "SELECT * FROM usrios_web_mnjo_flders_det WHERE email_usrio = '".$this->email."' AND nombre_folder = '".$row['nmbre_flder']."'";
+                $resf = $this->db->query($sqlf);
+                $file = array();
+                while($rowf = $resf->fetch(PDO::FETCH_ASSOC))
+                {
+                    array_push($file, $rowf['asnto_dcmnto'].'-'.$rowf['id_dcmnto']);
+                }
+                $folder[$row['nmbre_flder']] = $file;
+            }
+
+        }
+        $main['Bandeja de entrada'] = $folder;
+        return $main;
+
+    }
+
+    function del_doc($doc_id)
+    {
+        $sql = "DELETE FROM usrios_web_mnjo_flders_det WHERE email_usrio = '".$this->email."' AND id_dcmnto = ".$doc_id;
+        $res = $this->db->query($sql);
+        if($res)
+        {
+            return true;
+        }else{
+            return false;
+        }
+    }
+
+    function del_folder($fname)
+    {
+        $sqlfiles = "DELETE FROM usrios_web_mnjo_flders_det WHERE email_usrio = '".$this->email."' AND nombre_folder = '".$fname."'";
+        $resFiles = $this->db->query($sqlfiles);
+        if($resFiles)
+        {
+            $sql = "DELETE FROM usrios_web_mnjo_flders_enc WHERE email_usrio = '".$this->email."' AND nmbre_flder = '".$fname."'";
+            $res = $this->db->query($sql);
+            if($res)
+            {
+                return true;
+            }else{
+                return false;
+            }
+
+        }
+        
+
+    }
+    
+
